+++ nix/overrides.nix
@@ -1,7 +1,12 @@
 final: prev:
 
 let
-  inherit (final) lib;
+  inherit (final)
+    lib
+    stdenv
+    darwin
+    rustPlatform
+    ;
   utils = final.callPackage ./utils.nix { };
 
   # Mark broken packages here.
@@ -302,6 +307,31 @@ let
         withAllGrammars = withPlugins (_: allGrammars);
       };
     });
+
+    sniprun = super.sniprun.overrideAttrs (
+      old:
+      let
+        rust = rustPlatform.buildRustPackage {
+          pname = "sniprun-rust";
+          inherit (old) version src;
+          inherit (old.passthru.rust) cargoSha256;
+          buildInputs = lib.optionals stdenv.isDarwin [
+            darwin.apple_sdk.frameworks.Security
+          ];
+          doCheck = false;
+        };
+      in
+      {
+        postPatch = ''
+          sed -Ei lua/sniprun.lua -e '/local binary_path =/,+2c \
+          local binary_path = "${rust}/bin/sniprun"'
+        '';
+        propagatedBuildInputs = [ rust ];
+        passthru = (old.passthru or { }) // {
+          inherit rust;
+        };
+      }
+    );
   };
 in
 {